requirements-render.txt
Flask>=3.0.2
Flask-SocketIO>=5.0.0
gunicorn==21.2.0

Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.41
Flask-Migrate>=4.1.0
alembic>=1.16.1
psycopg2-binary>=2.9.10
redis>=5.0.0

Bootstrap-Flask>=2.3.3
Flask-Login>=0.6.3
Flask-CKEditor>=0.5.1

PyJWT>=2.10.1
bcrypt>=4.0.1
python-dotenv>=1.1.0
requests>=2.32.3
httpx>=0.28.0
sentry-sdk>=2.24.1
nest-asyncio>=1.6.0

langchain-core>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
google-genai>=0.2.0
langgraph>=0.4.5
langchain-mcp-adapters>=0.1.1
mcp>=1.9.0
tavily-python>=0.7.0
openai>=1.76.0
pydantic>=2.11.4
pydantic-settings>=2.9.1

google-api-python-client>=2.181.0
google-auth>=2.28.2
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.2

deepgram-sdk>=3.0.0
elevenlabs>=1.0.0
cartesia>=1.0.0
livekit==0.17.5
twilio>=8.0.0
numpy>=2.2.4
pandas>=2.2.3
scipy>=1.11.0
librosa>=0.10.0
soundfile>=0.12.1
websockets>=12.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
a2wsgi>=1.10.0
